refactor(smoke): give the two gates one shared assertion body
The wayland gate copied ~45 lines from the x11 one — run, assert success with
the stdout/stderr message, read the report, assert backend/mode/app.state/
app.value, run --dry-run, diff the rows — including two rationale comments that
are facts about `smoke` itself rather than about a backend.

common::assert_smoke_gate takes the backend and whatever environment the harness
itself needs (the x11 gate's private display), so each gate file now holds only
its own preamble.
