+++ crates/glass-mcp/tests/smoke_x11.rs
@@ -4,8 +4,7 @@
 
 mod common;
 
-use common::{Xvfb, assert_fixture_checks_pass, read_report, rows};
-use std::process::Command;
+use common::{Xvfb, assert_smoke_gate};
 
 const SERVER: &str = env!("CARGO_BIN_EXE_glass-mcp");
 
@@ -13,59 +12,5 @@ const SERVER: &str = env!("CARGO_BIN_EXE_glass-mcp");
 #[ignore = "requires Xvfb and a target app; run via: cargo test -p glass-mcp --test smoke_x11 -- --ignored"]
 fn smoke_x11_passes_against_a_real_app() {
     let xvfb = Xvfb::start();
-    let dir = tempfile::tempdir().expect("tempdir");
-    let report = dir.path().join("smoke.json");
-
-    let out = Command::new(SERVER)
-        .args(["smoke", "--backend", "x11", "--report"])
-        .arg(&report)
-        .env("DISPLAY", &xvfb.display)
-        .output()
-        .expect("run smoke");
-
-    let stdout = String::from_utf8_lossy(&out.stdout);
-    // A setup failure (no target app, no accessibility bus) writes no report and explains
-    // itself on stderr, so a stdout-only message would leave nothing to triage from.
-    let stderr = String::from_utf8_lossy(&out.stderr);
-    assert!(
-        out.status.success(),
-        "smoke failed:\n--- stdout ---\n{stdout}\n--- stderr ---\n{stderr}"
-    );
-
-    let json = read_report(&report);
-    assert_eq!(json["backend"], "x11");
-    assert_eq!(json["mode"], "full", "a real run is not a plan: {json}");
-    assert_eq!(
-        json["app"]["state"], "selected",
-        "a real run drove an app, so the report must say which: {json}"
-    );
-    assert!(
-        json["app"]["value"].as_str().is_some_and(|a| !a.is_empty()),
-        "the selected app must be named: {json}"
-    );
-
-    // The invariant `smoke/mod.rs` declares: a real run's rows are the `--dry-run` preview's
-    // rows. Sourcing the expectation from the same binary's own plan, rather than a list
-    // copied into this file, is what keeps the two from drifting apart unnoticed.
-    let plan_path = dir.path().join("plan.json");
-    let plan = Command::new(SERVER)
-        .args(["smoke", "--backend", "x11", "--dry-run", "--report"])
-        .arg(&plan_path)
-        .env("DISPLAY", &xvfb.display)
-        .output()
-        .expect("run smoke --dry-run");
-    assert!(
-        plan.status.success(),
-        "smoke --dry-run failed: {}",
-        String::from_utf8_lossy(&plan.stderr)
-    );
-    let plan = read_report(&plan_path);
-    assert_eq!(plan["mode"], "dry_run", "a plan must say so: {plan}");
-    assert_eq!(
-        rows(&json),
-        rows(&plan),
-        "a real run must carry exactly the checks --dry-run previews: {stdout}"
-    );
-
-    assert_fixture_checks_pass(&json, &stdout);
+    assert_smoke_gate(SERVER, "x11", &[("DISPLAY", xvfb.display.as_str())]);
 }
